Time: 16 ms (22.92%) | Memory: 31.9 MB (98.14%) - LeetSync

+++ 561-array-partition/array-partition.cpp
@@ -3,7 +3,7 @@ public:
     int arrayPairSum(vector<int>& nums) {
         int n=nums.size();
         sort (nums.begin(),nums.end());
-        int total=0;
+        long long total=0;
         for (int i=0;i<n;i+=2)
         {
             total+=nums[i];
